@@ -16,11 +16,10 @@
     "gulp-generate-ts-diagnostics": "^0.1.15",
     "gulp-rename": "^1.2.2",
     "mocha": "^2.2.5",
-    "morgan": "^1.6.1",
     "nightmare": "^1.8.2",
     "radium": "^0.13.2",
-    "react": "^0.13.3",
     "sinon": "^1.15.4",
+    "image-diff": "^1.2.0",
     "source-map-support": "^0.3.2",
     "typescript-register": "^1.1.0"
   },
@@ -32,12 +31,12 @@
     "connect-modrewrite": "^0.8.1",
     "es6-promise": "^2.3.0",
     "glob": "^5.0.13",
-    "gulp-generate-ts-diagnostics": "^0.1.16",
-    "image-diff": "^1.2.0",
-    "nightmare": "^1.8.2",
+    "morgan": "^1.6.1",
+    "react": "^0.14.0-beta1",
     "react-addons-pure-render-mixin": "^0.14.0-beta1",
+    "react-dom": "^0.14.0-beta1",
     "systemjs": "^0.18.4",
     "terminal-colors": "^0.1.3",
-    "typescript": "^1.5.0-beta"
+    "typescript": "https://github.com/Microsoft/TypeScript/tarball/01244d60dac82eec017e9150e663c5d144f989ff"
   }
 }
