@@ -2,12 +2,17 @@
 /// <reference path='../../../../../typings/react/react.d.ts'/>
 /// <reference path='../../../../../typings/react/react-jsx.d.ts'/>
 /// <reference path='../../../../../typings/radium/radium.d.ts'/>
+/// <reference path='../../../../../typings/platform/platform.d.ts' />
 /// <reference path='../../../../../typings/react-addons-pure-render-mixin/react-addons-pure-render-mixin.d.ts'/>
 
+let __r = require;
 import { ComposerContent } from '../../../../../src/client/components';
-import PureRenderMixin = require('react-addons-pure-render-mixin');
-import React = require('react');
-import Radium = require('radium');
+import ReactType = require('react');
+let React: typeof ReactType = inClient ? require('public/scripts/vendor/react') : __r('react');
+import PureRenderMixinType = require('react-addons-pure-render-mixin');
+let PureRenderMixin: typeof PureRenderMixinType = inClient ? require('/public/scripts/vendor/react-with-addons.js') : __r('react-addons-pure-render-mixin');
+import RadiumType = require('radium');
+let Radium: typeof RadiumType = inClient ? (window as any).Radium : __r('radium');
 
 interface NavigationBarProps {
     a: string;
@@ -16,6 +21,10 @@ interface NavigationBarProps {
 
 export class NavigationBar extends ComposerContent<NavigationBarProps, {}> {
     public mixins = [PureRenderMixin];
+
+    public componentDidMount() {
+        // console.log(React.findDOMNode(this));
+    }
 
     static fetch(): Promise<NavigationBarProps> {
         let promise = new Promise((resolve, reject) => {
