@@ -1,17 +1,17 @@
 
-interface RoutingInfo {
+interface Page {
     route: string;
-    document: EmitInfo;
-    layout: EmitInfo;
-    contents: EmitInfo[];
+    document: ComponentInfo;
+    layout: ComponentInfo;
+    contents: ComponentInfo[];
 }
 
-interface EmitInfo {
+interface ComponentInfo {
     className: string;
     importPath: string;
 }
 
 declare module ComposerRouter {
-    export function init(table: RoutingInfo[]): void;
+    export function init(table: Page[]): void;
     export function navigateTo(route: string): void;
 }
